+++ types/Forecast.ts
@@ -13,23 +13,24 @@ export type Recommendations = {
 	total: number // need to calculate on storage
 	consensus: string // need to calculate on storage
 	angle: number // the angle of the donut chart arrow, from 0-180
+	score: number
 	buy: number
 	hold: number
 	month: string // '2022-02-01',
 	sell: number
 	strongBuy: number
 	strongSell: number
-}[]
+}
 
 export type ForecastData = {
 	targets: Targets
-	recommendations: Recommendations
+	recommendations: Recommendations[]
 }
 
 export type Forecast = {
 	info: Info
 	data: {
 		targets: Targets
-		recommendations: Recommendations
+		recommendations: Recommendations[]
 	}
 }
